+++ core/tag/tags/LinkTag.php
@@ -39,7 +39,10 @@ class LinkTag extends BaseTag
      */
     public function readme()
     {
-        return 'Generate a link to an external page or an internal page url (target blank differenc). In order to call an internal URL use the `//` prefix like `link[//contact](Go to contact)` now the `//` are replace by the url of the webserver. In order to generate an external url use `link[luya.io](Go to Luya.io)`.';
+        return 'Generate a link to an external page or an internal page. 
+        In order to call an internal URL use the `//` prefix like `link[//contact](Go to contact)`, now the `//` are replace by the url of the webserver. 
+        If a single `/` is used its a relative url and therfore won\'t be changed.
+        In order to generate an external url use `link[luya.io](Go to Luya.io)`.';
     }
     
     /**
@@ -52,14 +55,18 @@ class LinkTag extends BaseTag
      */
     public function parse($value, $sub)
     {
-        if (substr($value, 0, 2) == '//') {
+        if (StringHelper::startsWith($value, '//')) {
+            // its an absolute url 
             $value = StringHelper::replaceFirst('//', Url::base(true) . '/', $value);
             $external = false;
+        } elseif (StringHelper::startsWith($value, '/')) {
+            // its a relative url, keep it like this
+            $external = false;
         } else {
+            $value = Url::ensureHttp($value);
             $external = true;
         }
         
-        $value = Url::ensureHttp($value);
         $label = empty($sub) ? $value : $sub;
         
         return Html::a($label, $value, ['class' => $external ? 'link-external' : 'link-internal', 'target' => $external ? '_blank' : null]);